divide signInWith* mutation test codes
